Improved p2p sync timings

+++ src/P2pNet.Base/ChannelPeers/PeerClockSyncInfo.cs
@@ -4,13 +4,13 @@ namespace P2pNet
     public class PeerClockSyncInfo
     {
         public string peerId;
-        public long networkLagMs; // round trip time / 2
+        public int networkLagMs; // round trip time / 2
         public double networkLagSigma; // Std deviation
-        public long sysClockOffsetMs; // localTime + offset = remoteTime
+        public int sysClockOffsetMs; // localTime + offset = remoteTime
         public double sysClockOffsetSigma;
-        public long msSinceLastSync;
+        public int msSinceLastSync;
         public long syncCount; // number of synd test samples
-        public PeerClockSyncInfo(string pid, long cnt, long since, long offset, double offsetSigma, long lag, double lagSigma)
+        public PeerClockSyncInfo(string pid, long cnt,int since, int offset, double offsetSigma, int lag, double lagSigma)
         {
             peerId = pid;
             syncCount = cnt;
